Moved Evaluator classes to their own file for better organization
