+++ llm_with_rag_chatbot/test_chatbot_conversation.py
@@ -24,116 +24,12 @@ import os
 import re
 
 from dotenv import load_dotenv
-from langchain_core.prompts import PromptTemplate
-from langchain_openai.chat_models import ChatOpenAI
-from rouge import Rouge
 
+from evaluators import KeywordEvaluator, RougelEvaluator, LlmEvaluator
 from llm_with_rag_chatbot.openai_chatbot_with_assistant_api import process
 
-TIME_FMT = "%Y-%M-%d-%H%M.%S.%f"
+TIME_FMT = "%Y-%m-%d-%H%M.%S.%f"
 ENC = "utf-8"
-
-class Evaluator:
-    """
-    Evaluates chatbot responses.
-    """
-    def __init__(self):
-        self._scores = []
-
-    def get_name(self) -> str:
-        """
-        Returns the name of the approach/metric used by this Evaluator.
-        """
-        pass
-
-    def score(self, response: str, exchange: dict) -> float:
-        """
-        Scores the chatbot response and makes an internal note of its correctness.
-
-        :param response: the chatbot response
-        :param exchange: a dict containing the query given to the chatbot and a rubric with which to evaluate the chatbot's response
-        """
-        pass
-
-    def record_response(self, response: str, exchange: dict) -> None:
-        """
-        Scores the chatbot response and makes an internal note of its correctness.
-
-        :param response: the chatbot response
-        :param exchange: a dict containing the query given to the chatbot and a rubric with which to evaluate the chatbot's response
-        """
-        self._scores.append(self.score(response, exchange))
-
-    def get_results(self) -> str:
-        """
-        Returns results of evaluating all chatbot responses.
-        """
-        points = sum(self._scores)
-        total = len(self._scores)
-        return "{} Evaluation Results:\nTotal score {}\n{} points out of {}\nIndividual items: {}".format(self.get_name(), points / total, points, total, self._scores)
-
-class KeywordEvaluator(Evaluator):
-    """
-    Evaluates chatbot responses based on the presence of keywords specified in the rubric.
-    See the file-level docstring for documentation of the "keywords" field of the rubric.
-    """
-    def score(self, response: str, exchange: dict) -> float:
-        keywords = exchange["rubric"]["keywords"]
-        for rule in keywords:
-            if rule == "containsAll":
-                if not all(term.casefold() in response.casefold() for term in keywords[rule]):
-                    return 0.0
-            elif rule == "containsAny":
-                if not any(term.casefold() in response.casefold() for term in keywords[rule]):
-                    return 0.0
-            else:
-                raise ValueError("Invalid field {} found in keywords!".format(rule))
-        return 1.0
-    
-    def get_name(self) -> str:
-        return "Keyword"
-
-class RougelEvaluator(Evaluator):
-    """
-    Evaluates each response by computing the ROUGE-L score between it and the standard.
-    """
-    def score(self, response: str, exchange: dict) -> float:
-        rouge = Rouge(metrics=["rouge-l"], stats=["f"])
-        return rouge.get_scores(response, exchange["rubric"]["standard"])[0]["rouge-l"]["f"]
-    
-    def get_name(self) -> str:
-        return "ROUGE-L"
-
-class LlmEvaluator(Evaluator):
-    """
-    Uses an LLM to assess how close the chatbot's responses are to gold standard responses.
-    """
-    def score(self, response: str, exchange: dict) -> float:
-        OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
-        model = ChatOpenAI(openai_api_key=OPENAI_API_KEY, model=os.getenv("MODEL"))
-        prompt = PromptTemplate.from_template(
-            """You are grading an agent's responses to questions.
-            Question: {question}
-            Example of correct response: {standard}
-            Agent's response: {response}
-            The agent's response is considered correct if it has the same meaning as the example response, even if it is not identical.
-            Is the agent's response correct?
-            Please answer CORRECT or INCORRECT.
-            """
-        )
-        fields = {
-            "question": exchange["query"],
-            "standard": exchange["rubric"]["standard"],
-            "response": response
-        }
-        result = model.invoke(prompt.invoke(fields))
-        if result.content.lower() == "correct":
-            return 1.0
-        else:
-            return 0.0
-        
-    def get_name(self) -> str:
-        return "LLM"
 
 def report(msg, f):
     """
@@ -144,6 +40,9 @@ def report(msg, f):
     f.write(msg)
 
 start_time = datetime.now()
+
+# Load environment variables
+load_dotenv()
 
 # Process command line arguments.
 parser = argparse.ArgumentParser()
@@ -160,7 +59,7 @@ args = parser.parse_args()
 conversation_history = deque(maxlen=10)
 
 # Keep track of correctness.
-evaluators = (KeywordEvaluator(), RougelEvaluator(), LlmEvaluator())
+evaluators = (KeywordEvaluator(), RougelEvaluator(), LlmEvaluator(os.getenv("OPENAI_API_KEY"), os.getenv("MODEL")))
 
 # Feed the chatbot each query in the conversation and score each resulting response.
 # Meanwhile, write the chatbot responses to the provided file for debugging purposes.
